MAINT: refactored utils functions to internals.parsing_helpers and remove deprecated utils module
