+++ fastf1/_api.py
@@ -6,7 +6,7 @@ import zlib
 import numpy as np
 import pandas as pd
 
-from fastf1._utils import (
+from fastf1.internals.parsing_helpers import (
     recursive_dict_get,
     to_datetime,
     to_timedelta
