+++ fastf1/livetiming/data.py
@@ -6,7 +6,7 @@ import json
 import warnings
 from datetime import timedelta
 
-from fastf1._utils import (
+from fastf1.internals.parsing_helpers import (
     recursive_dict_get,
     to_datetime
 )
