+++ fastf1/utils.py
@@ -9,9 +9,9 @@ release.
 import datetime
 import warnings
 
-from fastf1._utils import recursive_dict_get as _recursive_dict_get
-from fastf1._utils import to_datetime as _to_datetime
-from fastf1._utils import to_timedelta as _to_timedelta
+from fastf1.internals.parsing_helpers import recursive_dict_get as _recursive_dict_get
+from fastf1.internals.parsing_helpers import to_datetime as _to_datetime
+from fastf1.internals.parsing_helpers import to_timedelta as _to_timedelta
 
 
 _DEPRECATION_MSG = (
